+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated to Angular 9
 
 ### Added
+- Verses group visualization
 - Multiple line words normalization in interpretative and critical edition
 - Word visualization
 - Incurable corruptions visualization